Rename files to prevent awkward diff
